+++ Classes/Service/FlashMessageService.php
@@ -9,7 +9,7 @@ use \TYPO3\CMS\Core\Utility\GeneralUtility;
 use \T3v\T3vCore\Service\AbstractService;
 
 /**
- * Class FlashMessageService
+ * Flash Message Service Class
  *
  * @package T3v\T3vCore\Service
  */