+++ CHANGELOG.md
@@ -3,8 +3,8 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
-Unreleased
-----------
+1.2.0
+-----
 
 * Added `RendererUtility`
 * Added `MailService`